refactor: simplify results logic
