+++ src/pages/searchResultsPage/searchresultspage.scss
@@ -1,7 +1,7 @@
 // ------------------------------------------
 // Search Results Page
 // ------------------------------------------
-@use '../../../scss/partials' as *;
+@use '../../scss/partials' as *;
 
 // ------------------------------------------
 // Container
